services/python-llm/requirements.txt
fastapi
uvicorn
google-genai
pydantic
python-dotenv
beautifulsoup4